add safety to rm work

+++ cleanup-results-and-work.sh
@@ -4,6 +4,9 @@ shopt -s extglob
 
 rm -rf results
 
-cd work
-
-rm -rf  !(apptainer)
+if [ -d "work" ]; then
+	cd work
+	rm -rf  !(apptainer)
+else
+  echo "Work directory does not exist."
+fi
